services/validator/requirements.txt
kafka-python==2.0.2
jsonschema==4.20.0
PyYAML==6.0.1
python-json-logger==2.0.7
